refactor: msw common에서 getSpecialties로 네이밍 변경 #894
