+++ frontend/src/common/mock/getSpecialties/data.ts
