+++ frontend/src/common/mock/getSpecialties/handlers.ts
